+++ pymaid/fetch/__init__.py
@@ -56,6 +56,7 @@ from .. import core, utils, config, cache
 from navis import in_volume
 from .landmarks import get_landmarks, get_landmark_groups
 from .skeletons import get_skeleton_ids
+from .annnotations import get_annotation_graph
 
 
 __all__ = ['get_annotation_details', 'get_annotation_id',
@@ -1900,99 +1901,6 @@ def get_annotations(x, remote_instance=None):
         raise Exception(
             'No annotations retrieved. Make sure that the skeleton IDs exist.')
 
-
-def _entities_to_ann_graph(data, annotations_by_id=False, skeletons_by_id=True):
-    ann_ref = "id" if annotations_by_id else "name"
-    skel_ref = "id" if skeletons_by_id else "name"
-
-    g = nx.DiGraph()
-
-    for e in data["entities"]:
-        is_meta_ann = False
-
-        if e.get("type") == "neuron":
-            skids = e.get("skeleton_ids") or []
-            if len(skids) != 1:
-                logger.warning("Neuron with id %s is modelled by %s skeletons, ignoring", e["id"], len(skids))
-                continue
-            node_data = {
-                "name": e["name"],
-                "neuron_id": e["id"],
-                "is_skeleton": True,
-                "id": skids[0],
-            }
-            node_id = node_data[skel_ref]
-        else:  # is an annotation
-            node_data = {
-                "is_skeleton": False,
-                "id": e["id"],
-                "name": e["name"],
-            }
-            node_id = node_data[ann_ref]
-            is_meta_ann = True
-
-        anns = e.get("annotations", [])
-        if not anns:
-            g.add_node(node_id, **node_data)
-            continue
-
-        for ann in e.get("annotations", []):
-            g.add_edge(
-                ann[ann_ref],
-                node_id,
-                is_meta_annotation=is_meta_ann,
-            )
-
-        g.nodes[node_id].update(**node_data)
-
-    return g
-
-
-@cache.undo_on_error
-def get_annotation_graph(annotations_by_id=False, skeletons_by_id=True, remote_instance=None) -> nx.DiGraph:
-    """Get a networkx DiGraph of (meta)annotations and skeletons.
-
-    Can be slow for large projects.
-
-    Nodes in the graph have data:
-
-    Skeletons have
-    - id
-    - is_skeleton = True
-    - neuron_id (different to the skeleton ID)
-    - name
-
-    Annotations have
-    - id
-    - name
-    - is_skeleton = False
-
-    Edges in the graph have
-    - is_meta_annotation (whether it is between two annotations)
-
-    Parameters
-    ----------
-    annotations_by_id : bool, default False
-        Whether to index nodes representing annotations by their integer ID
-        (uses name by default)
-    skeletons_by_id : bool, default True
-        whether to index nodes representing skeletons by their integer ID
-        (True by default, otherwise uses the neuron name)
-    remote_instance : optional CatmaidInstance
-
-    Returns
-    -------
-    networkx.DiGraph
-    """
-    remote_instance = utils._eval_remote_instance(remote_instance)
-
-    query_url = remote_instance.make_url(remote_instance.project_id, "annotations", "query-targets")
-    post = {
-        "with_annotations": True,
-    }
-    data = remote_instance.fetch(query_url, post)
-
-    return _entities_to_ann_graph(data, annotations_by_id, skeletons_by_id)
 
 
 def filter_by_query(names: pd.Series, query: str, allow_partial: bool = False) -> pd.Series: